Update README, fix paths.
