+++ src/commands/cache/set.ts
@@ -13,36 +13,28 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { Command } from '@oclif/core';
-import { mapReplacer } from '../../../modules/helpers';
-import { ProjectCache } from '../../../modules/project-cache';
-import { ProjectConfig } from '../../../modules/project-config';
+import { ProjectCache } from '../../modules/project-cache';
+import { ProjectConfig } from '../../modules/project-config';
 
-export default class Get extends Command {
-    static description = 'Get the complete cache contents as JSON string or the value of a single key.';
+export default class Set extends Command {
+    static description = 'Set the cache value of an entry.';
 
-    static examples = [
-        `$ velocitas cache get
-{"foo":"bar"}`,
-        `$ velocitas cache get foo
-bar`,
+    static examples = [`$ velocitas cache set <key> <value>`];
+
+    static args = [
+        { name: 'key', description: 'The cache key to set', required: true },
+        { name: 'value', description: 'The value to set for the cache key', required: true },
     ];
 
-    static args = [{ name: 'key', description: 'The key of a single cache entry to get.', required: false }];
-
     async run(): Promise<void> {
-        const { args } = await this.parse(Get);
+        const { args } = await this.parse(Set);
 
         // although we are not reading the project config, we want to
         // ensure the command is run in a project directory only.
         ProjectConfig.read();
 
         const cache = ProjectCache.read();
-
-        let output = JSON.stringify(cache.raw(), mapReplacer);
-        if (args.key) {
-            output = cache.get(args.key);
-        }
-
-        this.log(output);
+        cache.set(args.key, args.value);
+        cache.write();
     }
 }
